@@ -1,7 +1,7 @@
 source 'https://rubygems.org'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.0.2'
+gem 'rails', '4.1.8'
 gem 'bootstrap-sass', '3.3.0.0'
 gem 'font-awesome-rails', '4.3.0.0'
 
@@ -51,11 +51,13 @@ group :development do
   gem 'mailcatcher'
 
   gem 'sqlite3'
+
+  gem 'annotate'
 end
 
 #Carga de Imagenes
 gem 'carrierwave'
 #gem 'rmagick'
 
-gem 'cancan'
+gem 'cancancan', '~> 1.10'
 
