@@ -3,11 +3,11 @@ class ApplicationController < ActionController::Base
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
 
-  before_filter do
-    resource = controller_name.singularize.to_sym
-    method = "#{resource}_params"
-    params[resource] &&= send(method) if respond_to?(method, true)
-  end
+  # before_action do
+  #   resource = controller_path.singularize.gsub('/', '_').to_sym
+  #   method = "#{resource}_params"
+  #   params[resource] &&= send(method) if respond_to?(method, true)
+  # end
 
   rescue_from CanCan::AccessDenied do |exception|
     flash[:alert] = 'Usted no esta autorizado para acceder a la página solicitada.'
@@ -27,6 +27,14 @@ class ApplicationController < ActionController::Base
       @current_permissions = current_user.roles.each do |role|
         role.permissions.collect{|i| [i.subject_class, i.action]}
       end
+    end
+
+    def set_user
+      @user = User.find_by(id: ActionController::Parameters.new(id: params[:id]).permit(:id)[:id])
+    end
+
+    def user_params
+      params.require(:user).permit(:name, :username, :email, :password, :locked, :mall_id, :cellphone, :role_id)#, rol_ids: [])
     end
 
 end
